Move helpers to lib directory
